+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog #
 
+## v1.1.0 (2015-01-12) ##
+
+  * Improvements in code quality and documentation
+  * Added UrlInfo::getDefaultPort()
+  * Added UrlInfo::getFileExtension()
+
 ## v1.0.3 (2014-06-01) ##
 
   * Code cleanup and documentation fixes